pyproject.toml
[project]
name = "cloud-storage-mixpanel"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "AK", email = "ak@mixpanel.com"},
]
dependencies = [
    "google-cloud-storage>=2.17.0",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
